refactor: extract EditExerciseModal, TagManagerModal, ExercisePickerModal (#745)
Each modal is now a props+emits child owning only its transient UI state
(form fields, rename/expand/add modes, focus trap); WorkoutTracker keeps
ownership of which modal is open, all store mutations, undo toasts, and
analytics. DOM, classes, and aria-labelledby ids are unchanged.

Also hoists the two helpers the modals shared with the log-set form into
lib: inputLimits (MAX_WEIGHT/MAX_REPS) and keyboardViewport
(scrollInputAboveKeyboard).

WorkoutTracker: ~2,880 -> ~2,450 lines.
