+++ src/components/WorkoutTracker.vue
@@ -616,233 +616,37 @@
   </Teleport>
 
 
-  <!-- Edit Exercise Modal -->
-  <Teleport to="body">
-    <div v-if="editTarget !== null" class="repMaxOverlay" @click.self="editTarget = null" @keydown.escape="editTarget = null">
-      <div class="repMaxModal" role="dialog" aria-modal="true" aria-labelledby="edit-exercise-title">
-        <h2 id="edit-exercise-title">Edit Exercise</h2>
-        <label class="repMaxLabel">
-          Name
-          <div class="repMaxInputRow">
-            <input
-              v-model.trim="editName"
-              type="text"
-              class="repMaxInput"
-              autocomplete="off"
-              maxlength="50"
-            />
-          </div>
-        </label>
-        <div class="repMaxLabel">
-          Tags
-          <div class="wtTagPicker">
-            <button
-              v-for="tag in availableEditTags"
-              :key="tag"
-              :class="['wtTagPickerChip', { wtTagPickerChipActive: editTags.includes(tag) }]"
-              :style="!editTags.includes(tag)
-                ? { borderColor: 'var(--border-strong)', color: 'var(--text-secondary)' }
-                : {}"
-              @click="toggleEditTag(tag)"
-            >{{ tag }}</button>
-            <span v-if="editTagAdding" class="wtTagInlineAdd">
-              <input
-                v-model.trim="newTagInput"
-                type="text"
-                autocomplete="off"
-                placeholder="Tag name"
-                maxlength="30"
-                class="wtTagInlineInput"
-                aria-label="New tag name"
-                ref="editTagInputEl"
-                @keyup.enter="addEditTag"
-                @blur="finishEditTagAdd"
-              />
-            </span>
-            <button v-else class="wtTagPickerChip wtTagAddChip" @mousedown.prevent @click="startEditTagAdd" aria-label="Add tag">+</button>
-          </div>
-        </div>
-        <!-- Plate calculator settings (iOS grouped style) -->
-        <div class="iosSettingsSection">
-          <span class="iosSettingsHeader">Input Mode</span>
-          <div class="iosSettingsGroup">
-            <div class="iosSettingsRow">
-              <span class="iosSettingsRowLabel">Plate calculator</span>
-              <button
-                class="iosToggle"
-                :class="{ iosToggleOn: editPlateMode }"
-                role="switch"
-                :aria-checked="editPlateMode"
-                @click="editPlateMode = !editPlateMode"
-              >
-                <span class="iosToggleKnob"></span>
-              </button>
-            </div>
-            <template v-if="editPlateMode">
-              <div class="iosSettingsRow">
-                <span class="iosSettingsRowLabel">Counting</span>
-                <div class="iosSegmentedControl">
-                  <button
-                    :class="['iosSegment', { iosSegmentActive: editPlateCountMode === 'per-side' }]"
-                    @click="editPlateCountMode = 'per-side'"
-                  >Per side</button>
-                  <button
-                    :class="['iosSegment', { iosSegmentActive: editPlateCountMode === 'total' }]"
-                    @click="editPlateCountMode = 'total'"
-                  >Total</button>
-                </div>
-              </div>
-              <div class="iosSettingsRow">
-                <span class="iosSettingsRowLabel">Starting weight</span>
-                <div class="iosStepper">
-                  <button class="iosStepperBtn" @click="editBarWeight = Math.max(0, editBarWeight - 5)" aria-label="Decrease weight">−</button>
-                  <input
-                    v-if="editBarWeightEditing"
-                    ref="editBarWeightInputEl"
-                    :value="editBarWeight"
-                    type="text"
-                    inputmode="numeric"
-                    autocomplete="off"
-                    class="iosStepperInput"
-                    aria-label="Starting weight"
-                    @focus="($event.target as HTMLInputElement)?.select(); scrollInputAboveKeyboard($event.target as HTMLElement)"
-                    @blur="editBarWeight = Math.max(0, Math.min(MAX_WEIGHT, Math.round(Number(($event.target as HTMLInputElement).value) || 0))); editBarWeightEditing = false"
-                  />
-                  <button v-else class="iosStepperValue iosStepperValueTappable" @click="editBarWeightEditing = true; nextTick(() => editBarWeightInputEl?.focus())">{{ editBarWeight }} {{ weightUnit }}</button>
-                  <button class="iosStepperBtn" @click="editBarWeight = Math.min(MAX_WEIGHT, editBarWeight + 5)" aria-label="Increase weight">+</button>
-                </div>
-              </div>
-            </template>
-          </div>
-        </div>
-        <div class="repMaxActions">
-          <button class="repMaxBtn repMaxBtnCalc" :disabled="!editName" @click="confirmEditExercise">Save</button>
-          <button class="repMaxBtn repMaxBtnClose" @click="editTarget = null">Cancel</button>
-        </div>
-        <button
-          v-if="editTargetIsArchived"
-          class="wtEditArchiveBtn"
-          @click="handleUnarchiveFromEdit"
-        >Unarchive Exercise</button>
-        <button
-          v-else
-          class="wtEditArchiveBtn"
-          @click="handleArchiveFromEdit"
-        >Archive Exercise</button>
-        <p class="wtEditArchiveHint">Hides this exercise from the main list — sets and PRs are preserved.</p>
-        <button
-          v-if="!confirmDeleteExercise"
-          class="wtEditDeleteBtn"
-          @click="confirmDeleteExercise = true"
-          aria-label="Delete exercise"
-        >Delete Exercise</button>
-        <div v-else class="wtEditDeleteConfirm">
-          <span class="wtEditDeleteConfirmText">Delete this exercise and all its sets?</span>
-          <div class="wtEditDeleteConfirmActions">
-            <button class="wtEditDeleteConfirmBtn wtEditDeleteConfirmCancel" @click="confirmDeleteExercise = false">Cancel</button>
-            <button class="wtEditDeleteConfirmBtn wtEditDeleteConfirmDanger" @click="undoDeleteExercise(store.exercises.find(e => e.id === editTarget)!); editTarget = null">Delete</button>
-          </div>
-        </div>
-      </div>
-    </div>
-  </Teleport>
+  <!-- Edit Exercise Modal (extracted to EditExerciseModal.vue) -->
+  <EditExerciseModal
+    :exercise="editTargetExercise"
+    :all-tags="store.allTags"
+    @close="editTarget = null"
+    @save="onEditExerciseSave"
+    @archive="handleArchiveFromEdit"
+    @unarchive="handleUnarchiveFromEdit"
+    @delete="onEditExerciseDelete"
+  />
 
-  <!-- Exercise Picker (timeline + Log Set) -->
-  <Teleport to="body">
-    <div v-if="timelineLogPicking" class="repMaxOverlay" @click.self="timelineLogPicking = false" @keydown.escape="timelineLogPicking = false">
-      <div class="repMaxModal" role="dialog" aria-modal="true" aria-labelledby="timeline-picker-title">
-        <h2 id="timeline-picker-title">Choose Exercise</h2>
-        <div class="wtExPickerList">
-          <button
-            v-for="ex in store.activeExercises"
-            :key="ex.id"
-            class="wtExPickerRow"
-            @click="pickExerciseForLog(ex.id)"
-          >
-            <span class="wtExPickerName">{{ ex.name }}</span>
-            <span class="wtChevron">›</span>
-          </button>
-          <button
-            class="wtExPickerRow wtExPickerNew"
-            @click="pickNewExerciseFromPicker"
-          >
-            <span class="wtExPickerName">+ New exercise</span>
-            <span class="wtChevron">›</span>
-          </button>
-        </div>
-        <div class="repMaxActions">
-          <button class="repMaxBtn repMaxBtnClose" @click="timelineLogPicking = false">Cancel</button>
-        </div>
-      </div>
-    </div>
-  </Teleport>
+  <!-- Exercise Picker (timeline + Log Set; extracted to ExercisePickerModal.vue) -->
+  <ExercisePickerModal
+    :open="timelineLogPicking"
+    :exercises="store.activeExercises"
+    @close="timelineLogPicking = false"
+    @select="pickExerciseForLog"
+    @create-new="pickNewExerciseFromPicker"
+  />
 
-  <!-- Tag Manager Modal -->
-  <Teleport to="body">
-    <div v-if="tagManagerOpen" class="repMaxOverlay" @click.self="tagManagerOpen = false" @keydown.escape="tagManagerOpen = false">
-      <div class="repMaxModal" role="dialog" aria-modal="true" aria-labelledby="tag-manager-title">
-        <h2 id="tag-manager-title">Manage Tags</h2>
-        <p v-if="store.allTags.length === 0 && !tagManagerAdding" class="wtEmpty" style="margin: var(--space-4) 0">No tags yet. Tap + to create one.</p>
-        <ul class="wtTagManagerList">
-          <li v-for="tag in store.allTags" :key="tag" class="wtTagManagerItemWrap">
-            <div class="wtTagManagerItem">
-              <template v-if="renamingTag === tag">
-                <input
-                  v-model.trim="renameTagValue"
-                  type="text"
-                  autocomplete="off"
-                  maxlength="30"
-                  class="repMaxInput wtTagManagerInput"
-                  aria-label="Rename tag"
-                  @keyup.enter="confirmRenameTag"
-                  @keyup.escape="renamingTag = null"
-                  ref="renameTagInputEl"
-                />
-                <button class="wtTagManagerSaveBtn" @click="confirmRenameTag" :disabled="!renameTagValue" aria-label="Save tag name">✓</button>
-                <button class="wtTagManagerCancelBtn" @click="renamingTag = null" aria-label="Cancel rename">✕</button>
-              </template>
-              <template v-else>
-                <button class="wtTagManagerExpandBtn" @click="toggleTagExpand(tag)" :aria-expanded="expandedTag === tag" :aria-label="'Show exercises for ' + tag">
-                  <span class="wtTagManagerExpandIcon" :class="{ expanded: expandedTag === tag }">›</span>
-                </button>
-                <span class="wtTagManagerLabel" @click="toggleTagExpand(tag)" role="button" tabindex="0" @keydown.enter="toggleTagExpand(tag)" @keydown.space.prevent="toggleTagExpand(tag)">{{ tag }}</span>
-                <span class="wtTagManagerCount">{{ tagExerciseCount(tag) }}</span>
-                <button class="wtTagManagerEditBtn" @click="startRenameTag(tag)" aria-label="Rename tag">✎</button>
-                <button class="wtTagManagerDeleteBtn" @click="confirmDeleteTag(tag)" aria-label="Delete tag">✕</button>
-              </template>
-            </div>
-            <ul v-if="expandedTag === tag" class="wtTagExerciseList">
-                <li v-for="exercise in store.exercises" :key="exercise.id">
-                  <button class="wtTagExerciseRow" @click="toggleExerciseTag(exercise.id, tag)">
-                    <span class="wtTagExerciseRowName">{{ exercise.name }}</span>
-                    <svg v-if="exercise.tags.includes(tag)" class="wtTagExerciseCheck" viewBox="0 0 24 24" fill="none" stroke="var(--accent)" stroke-width="2.5" stroke-linecap="round" stroke-linejoin="round" width="20" height="20"><polyline points="20 6 9 17 4 12"/></svg>
-                  </button>
-                </li>
-              </ul>
-          </li>
-        </ul>
-        <div v-if="tagManagerAdding" class="wtTagManagerAddRow">
-          <input
-            v-model.trim="tagManagerNewName"
-            type="text"
-            autocomplete="off"
-            placeholder="Tag name"
-            maxlength="30"
-            class="repMaxInput"
-            aria-label="New tag name"
-            ref="tagManagerInputEl"
-            @keyup.enter="confirmTagManagerAdd"
-            @keyup.escape="cancelTagManagerAdd"
-          />
-          <button class="wtTagAddBtn" @mousedown.prevent @click="confirmTagManagerAdd" :disabled="!tagManagerNewName" aria-label="Create tag">✓</button>
-        </div>
-        <div class="repMaxActions">
-          <button v-if="!tagManagerAdding" class="repMaxBtn repMaxBtnCalc" @click="startTagManagerAdd">+ New Tag</button>
-          <button class="repMaxBtn repMaxBtnClose" @click="tagManagerOpen = false">Done</button>
-        </div>
-      </div>
-    </div>
-  </Teleport>
+  <!-- Tag Manager Modal (extracted to TagManagerModal.vue) -->
+  <TagManagerModal
+    :open="tagManagerOpen"
+    :all-tags="store.allTags"
+    :exercises="store.exercises"
+    @close="tagManagerOpen = false"
+    @create-tag="store.addCustomTag"
+    @rename-tag="onRenameTag"
+    @delete-tag="confirmDeleteTag"
+    @toggle-exercise-tag="toggleExerciseTag"
+  />
 
   <!-- Rest timer bar -->
   <button
@@ -900,6 +704,11 @@ import { computeWeeklyGoal } from '../lib/weeklyGoal'
 import ExerciseDetailModal from '../views/ExerciseDetailModal.vue'
 import RestTimerContent from './RestTimerContent.vue'
 import WorkoutTimeline from './WorkoutTimeline.vue'
+import EditExerciseModal, { type EditExerciseSave } from './EditExerciseModal.vue'
+import TagManagerModal from './TagManagerModal.vue'
+import ExercisePickerModal from './ExercisePickerModal.vue'
+import { scrollInputAboveKeyboard } from '../lib/keyboardViewport'
+import { MAX_WEIGHT, MAX_REPS } from '../lib/inputLimits'
 
 const store = useWorkoutStore()
 const progressionStore = useProgressionStore()
@@ -1202,8 +1011,6 @@ watch(() => store.allTags, (tags) => {
 const detailExerciseId = ref<string | null>(null)
 
 const logModalFocus = useFocusTrap()
-const editExerciseFocus = useFocusTrap()
-const tagManagerFocus = useFocusTrap()
 
 // ── Swipe-to-dismiss for log-set sheet (step 5f) ────────────────
 // Drag the handle (or the sheet body, when not scrolled) down past
@@ -2427,35 +2234,6 @@ const bestWeightAtReps = computed<number | null>(() => {
   return best > 0 ? best : null
 })
 
-const MAX_WEIGHT = 2000
-
-/** Shrink modal to fit above iOS keyboard, then scroll input into view */
-function scrollInputAboveKeyboard(el: HTMLElement) {
-  setTimeout(() => {
-    const modal = el.closest('.repMaxModal') as HTMLElement | null
-    if (!modal) return
-    const vv = window.visualViewport
-    if (!vv) return
-    // Shrink modal so it fits within the visible viewport above the keyboard
-    const availableHeight = vv.height - 96
-    modal.style.maxHeight = `${availableHeight}px`
-    // Scroll the input into view within the now-scrollable modal
-    nextTick(() => {
-      const inputRect = el.getBoundingClientRect()
-      const visibleBottom = vv.offsetTop + vv.height
-      if (inputRect.bottom > visibleBottom - 16) {
-        modal.scrollTop += inputRect.bottom - visibleBottom + 60
-      }
-    })
-    // Restore max-height when keyboard dismisses
-    const restore = () => {
-      modal.style.maxHeight = ''
-      vv.removeEventListener('resize', restore)
-    }
-    vv.addEventListener('resize', restore)
-  }, 400)
-}
-const MAX_REPS = 200
 const hasSetData = computed(() => weight.value !== null && weight.value > 0 && weight.value <= MAX_WEIGHT && reps.value !== null && reps.value >= 1 && reps.value <= MAX_REPS)
 
 const canSave = computed(() => {
@@ -2631,12 +2409,6 @@ function undoDeleteExercise(exercise: Exercise) {
 const archivedOpen = ref(false)
 const archivedListId = 'wt-archived-list'
 
-const editTargetIsArchived = computed(() => {
-  if (!editTarget.value) return false
-  const ex = store.exercises.find(e => e.id === editTarget.value)
-  return !!ex?.archived_at
-})
-
 function handleArchiveFromEdit() {
   const id = editTarget.value
   if (!id) return
@@ -2668,90 +2440,33 @@ function unarchiveExerciseFromList(exerciseId: string) {
   logEvent('exercise_unarchive')
 }
 
-// ── Edit exercise state (rename + tags) ──────────────────────────
+// ── Edit exercise modal (extracted to EditExerciseModal.vue) ─────
+// The parent owns which exercise is being edited and applies the saved
+// form to the store; the modal owns the transient form state.
 const editTarget = ref<string | null>(null)
-const confirmDeleteExercise = ref(false)
-const editName = ref('')
-const editTags = ref<string[]>([])
-const newTagInput = ref('')
-const editPlateMode = ref(false)
-const editPlateCountMode = ref<'per-side' | 'total'>('per-side')
-const editBarWeight = ref<number>(45)
-const editBarWeightEditing = ref(false)
-const editBarWeightInputEl = ref<HTMLInputElement | null>(null)
 
+const editTargetExercise = computed<Exercise | null>(() =>
+  store.exercises.find(e => e.id === editTarget.value) ?? null
+)
 
 function openEditExerciseModal(exercise: Exercise) {
   editTarget.value = exercise.id
-  confirmDeleteExercise.value = false
-  editName.value = exercise.name
-  editTags.value = [...(exercise.tags || [])]
-  editPlateMode.value = exercise.inputMode === 'plates'
-  editPlateCountMode.value = exercise.plateCountMode || 'per-side'
-  editBarWeight.value = exercise.barWeight ?? (exercise.plateCountMode === 'total' ? 0 : 45)
-  newTagInput.value = ''
 }
 
-const editTagInputEl = ref<HTMLInputElement | null>(null)
-const editTagAdding = ref(false)
-
-function startEditTagAdd() {
-  editTagAdding.value = true
-  nextTick(() => editTagInputEl.value?.focus())
-}
-
-function addEditTag() {
-  const tag = newTagInput.value.trim()
-  if (tag && !editTags.value.includes(tag)) {
-    editTags.value.push(tag)
-  }
-  newTagInput.value = ''
-  nextTick(() => editTagInputEl.value?.focus())
-}
-
-function finishEditTagAdd() {
-  const tag = newTagInput.value.trim()
-  if (tag && !editTags.value.includes(tag)) {
-    editTags.value.push(tag)
-  }
-  newTagInput.value = ''
-  editTagAdding.value = false
-}
-
-
-function toggleEditTag(tag: string) {
-  if (editTags.value.includes(tag)) {
-    editTags.value = editTags.value.filter(t => t !== tag)
-  } else {
-    editTags.value.push(tag)
-  }
-}
-
-// All known tags, including any on this exercise that might not be in allTags yet
-const availableEditTags = computed(() => {
-  const all = new Set([...store.allTags, ...editTags.value])
-  return [...all]
-})
-
-function confirmEditExercise() {
-  if (!editTarget.value || !editName.value) return
-  // Auto-add any pending tag text
-  const pendingTag = newTagInput.value.trim()
-  if (pendingTag && !editTags.value.includes(pendingTag)) {
-    editTags.value.push(pendingTag)
-  }
-  store.renameExercise(editTarget.value, editName.value)
-  store.updateExerciseTags(editTarget.value, editTags.value)
+function onEditExerciseSave(payload: EditExerciseSave) {
+  if (!editTarget.value) return
+  store.renameExercise(editTarget.value, payload.name)
+  store.updateExerciseTags(editTarget.value, payload.tags)
   // Save input mode and plate settings
-  store.setExerciseInputMode(editTarget.value, editPlateMode.value ? 'plates' : 'numpad')
-  if (editPlateMode.value) {
-    store.setExercisePlateCountMode(editTarget.value, editPlateCountMode.value)
-    store.setExerciseBarWeight(editTarget.value, editBarWeight.value)
+  store.setExerciseInputMode(editTarget.value, payload.plateMode ? 'plates' : 'numpad')
+  if (payload.plateMode) {
+    store.setExercisePlateCountMode(editTarget.value, payload.plateCountMode)
+    store.setExerciseBarWeight(editTarget.value, payload.barWeight)
   }
   editTarget.value = null
   // When switching to plate mode, reverse-sync the current weight into
   // plates so the user's entered value is preserved (LIFT-388 review fix).
-  if (editPlateMode.value && weight.value) {
+  if (payload.plateMode && weight.value) {
     syncPlatesFromWeight()
   } else {
     syncPlateWeight()
@@ -2759,45 +2474,18 @@ function confirmEditExercise() {
   logEvent('exercise_edit')
 }
 
-// ── Tag manager ────────────────────────────────────────────────
+function onEditExerciseDelete() {
+  const exercise = editTargetExercise.value
+  if (!exercise) return
+  undoDeleteExercise(exercise)
+  editTarget.value = null
+}
+
+// ── Tag manager (extracted to TagManagerModal.vue) ─────────────
 const tagManagerOpen = ref(false)
-const renamingTag = ref<string | null>(null)
-const renameTagValue = ref('')
-const renameTagInputEl = ref<HTMLInputElement[] | null>(null)
-const expandedTag = ref<string | null>(null)
-const tagManagerAdding = ref(false)
-const tagManagerNewName = ref('')
-const tagManagerInputEl = ref<HTMLInputElement | null>(null)
+
 function openTagManager() {
   tagManagerOpen.value = true
-  renamingTag.value = null
-  expandedTag.value = null
-  tagManagerAdding.value = false
-  tagManagerNewName.value = ''
-}
-
-function startTagManagerAdd() {
-  tagManagerAdding.value = true
-  nextTick(() => tagManagerInputEl.value?.focus())
-}
-
-function confirmTagManagerAdd() {
-  const tag = tagManagerNewName.value.trim()
-  if (tag && !store.allTags.includes(tag)) {
-    store.addCustomTag(tag)
-    expandedTag.value = tag
-  }
-  tagManagerNewName.value = ''
-  tagManagerAdding.value = false
-}
-
-function cancelTagManagerAdd() {
-  tagManagerNewName.value = ''
-  tagManagerAdding.value = false
-}
-
-function toggleTagExpand(tag: string) {
-  expandedTag.value = expandedTag.value === tag ? null : tag
 }
 
 function toggleExerciseTag(exerciseId: string, tag: string) {
@@ -2810,34 +2498,17 @@ function toggleExerciseTag(exerciseId: string, tag: string) {
   store.updateExerciseTags(exerciseId, newTags)
 }
 
-function tagExerciseCount(tag: string): number {
-  return store.exercises.filter(e => (e.tags || []).includes(tag)).length
-}
-
-function startRenameTag(tag: string) {
-  renamingTag.value = tag
-  renameTagValue.value = tag
-  nextTick(() => {
-    if (renameTagInputEl.value && renameTagInputEl.value.length > 0) {
-      renameTagInputEl.value[0].focus()
-      renameTagInputEl.value[0].select()
-    }
-  })
-}
-
-function confirmRenameTag() {
-  if (!renamingTag.value || !renameTagValue.value) return
-  store.renameTag(renamingTag.value, renameTagValue.value)
+function onRenameTag(oldName: string, newName: string) {
+  store.renameTag(oldName, newName)
   logEvent('tag_rename')
-  renamingTag.value = null
 }
 
 function confirmDeleteTag(tag: string) {
-  const count = tagExerciseCount(tag)
   // Track which exercises have this tag for undo
   const affectedIds = store.exercises
     .filter(e => (e.tags || []).includes(tag))
     .map(e => e.id)
+  const count = affectedIds.length
   store.deleteTag(tag)
   logEvent('tag_delete')
   showUndo(
@@ -2871,30 +2542,6 @@ watch(showModal, async (open) => {
   } else {
     logModalFocus.deactivate()
     logSwipe.detach()
-  }
-})
-
-watch(editTarget, async (target) => {
-  if (target) {
-    await nextTick()
-    const el = document.querySelector<HTMLElement>('[aria-labelledby="edit-exercise-title"]')
-    if (el) {
-      editExerciseFocus.activate(el)
-      // Don't auto-focus the name input — user usually isn't renaming
-      ;(document.activeElement as HTMLElement)?.blur()
-    }
-  } else {
-    editExerciseFocus.deactivate()
-  }
-})
-
-watch(tagManagerOpen, async (open) => {
-  if (open) {
-    await nextTick()
-    const el = document.querySelector<HTMLElement>('[aria-labelledby="tag-manager-title"]')
-    if (el) tagManagerFocus.activate(el)
-  } else {
-    tagManagerFocus.deactivate()
   }
 })
