+++ src/lib/inputLimits.ts
@@ -1,0 +1,7 @@
+/**
+ * Validation bounds for weight/rep inputs, shared by the log-set modal and
+ * the exercise settings forms. Values above these are treated as typos, not
+ * ambition.
+ */
+export const MAX_WEIGHT = 2000
+export const MAX_REPS = 200
